Adding a separator to ALIASES
In some cases it is very inconvenient to have a comma as argument separator in an alias (e.g. when using function definitions or templates) as all internal commas have to be escaped.
By introducing a separator (like e.g. `;`) this can be made more convenient, easier to maintain and more understandable.

Also placed software for alias handling in a separate file.
