+++ doc/custcmd.doc
@@ -107,6 +107,15 @@ i.e.
 \endverbatim
 given the alias definition of `\l` in the example above.
 
+By default the separator for arguments in an alias ist the comma though when having e.g templates of function
+definitions as an argument this will lead to problems. It is also possible to specify a separator  with an alias
+in its definition like:
+\verbatim
+ALIASES += xreflist{3,;}="\xrefitem \1 \"\2\" \"\3\" "
+ALIASES += reminder="\xreflist{reminders;Reminder;Reminders}"
+\endverbatim
+
+
 \section custcmd_nesting Nesting custom command
 
 You can use commands as arguments of aliases, including commands
